Moved content and type counts queries to centralised class
